chore(refactor): move command arg definitions to commands module
