--- gitopscli/cli/__init__.py
@@ -1,11 +1,0 @@
-from .command_args import (
-    CommandArgs,
-    DeployArgs,
-    SyncAppsArgs,
-    AddPrCommentArgs,
-    CreatePreviewArgs,
-    CreatePrPreviewArgs,
-    DeletePreviewArgs,
-    DeletePrPreviewArgs,
-    VersionArgs,
-)
