+++ gitopscli/commands/__init__.py
@@ -1,1 +1,10 @@
-from .base import Command, CommandFactory
+from .command import Command, CommandArgs, CommandFactory
+
+from .add_pr_comment import AddPrCommentArgs
+from .create_preview import CreatePreviewArgs
+from .create_pr_preview import CreatePrPreviewArgs
+from .delete_preview import DeletePreviewArgs
+from .delete_pr_preview import DeletePrPreviewArgs
+from .deploy import DeployArgs
+from .sync_apps import SyncAppsArgs
+from .version import VersionArgs
