--- gitopscli/commands/base/__init__.py
@@ -1,2 +1,0 @@
-from .command import Command
-from .command_factory import CommandFactory
