--- gitopscli/commands/base/command.py
@@ -1,7 +1,0 @@
-from abc import ABCMeta, abstractmethod
-
-
-class Command(metaclass=ABCMeta):
-    @abstractmethod
-    def execute(self) -> None:
-        raise NotImplementedError
